// src/engine/client/types.ts
type Brand<B> = { [__brand]: B };
type Branded<T, B> = T & Brand<B>;

export enum Player {
  A = "A",
  B = "B",
}

export enum CardClass {
  POKEMON = "pokemon",
  ITEM = "item",
  SUPPORTER = "supporter",
  TOOL = "tool",
}

export enum EnergyType {
  GRASS = "grass",
  FIRE = "fire",
  WATER = "water",
  LIGHTNING = "lightning",
  FIGHTING = "fighting",
  PSYCHIC = "psychic",
  DARK = "dark",
  METAL = "metal",
}

export enum PokemonType {
  GRASS = "grass",
  FIRE = "fire",
  WATER = "water",
  LIGHTNING = "lightning",
  FIGHTING = "fighting",
  PSYCHIC = "psychic",
  DARK = "dark",
  METAL = "metal",
  COLORLESS = "colorless",
  DRAGON = "dragon",
}

export type CardGameId = Branded<string, "CardGameId">;
export type CardStableId = Branded<string, "CardStableId">;

export type CardReference<T extends CardClass> = {
  cardId: CardGameId;
  cardStableId: CardStableId;
  cardClass: T;
};

export type EnergyCount = number;
export type HealthPoints = number;
// ...
export type AttackId = Branded<string, "AttackId">;
// ...
export type BaseCardConfig = {
  stableId: CardStableId;
  cardClass: CardClass;
  name: string;
};

export type SpeciesId = Branded<string, "SpeciesId">;

export enum PokemonStage {
  BASIC = "basic",
  STAGE_1 = "stage_1",
  STAGE_2 = "stage_2",
}

export type PokemonCardConfig = BaseCardConfig & {
  cardClass: CardClass.POKEMON;
  speciesId: SpeciesId;
  type: PokemonType;
  baseHealthPoints: HealthPoints;
  evolvesFromSpeciesId: SpeciesId | null;
  stage: PokemonStage;
  isEx: boolean;
  isMegaEx: boolean;
  isFossil: boolean;
  attacks: AttackId[];
  typeWeaknesses: PokemonType[];
  retreatCost: EnergyCount;
};